+++ src/classes/general/status/Currency.js
@@ -1,7 +1,11 @@
 class Currency {
   constructor ({name, symbol, value = 0}) {
-    this.value = value
+    this.setValue(value)
     Object.assign(this, {name, symbol})
+  }
+  setValue (value) {
+    if (typeof value !== 'number') throw new Error('Currency value must be a number')
+    this.value = value
   }
   getValue () {
     return this.value