+++ src/spec/classes/shop.test.js
@@ -1,4 +1,4 @@
-import { gg, factory, error} from '../helpers'
+import { gg, factory, error } from '../helpers'
 
 let shop = factory.shop()
 let hero = factory.hero()
@@ -9,14 +9,14 @@ let goldCurrency = factory.goldCurrency()
 
 describe('shop', () => {
   test('should be able add items', () => {
-    shop.addItem(sword)
-    shop.addItem(dagger)
+    shop.addItem({item: sword, price: goldCurrency})
+    shop.addItem({item: dagger, price: goldCurrency})
     let shopItems = shop.getItems()
     expect(shopItems[0].item.name).toEqual(sword.name)
     expect(shopItems[1].item.name).toEqual(dagger.name)
   })
   test('should stack existent items', () => {
-    shop.addItem(sword)
+    shop.addItem({item: sword, price: goldCurrency})
     let swords = shop.getItems().filter(shopItem => shopItem.item.name === 'Great sword')[0].quantity
     expect(swords).toBe(2)
   })
@@ -61,15 +61,17 @@ describe('shop', () => {
     return expect(hero.interact(shop).sell(heroSword)).rejects.toEqual(new Error('Character must have the item to sell'))
   })
   test('should be able to buy item', () => {
-    shop.addItem(sword)
+    goldCurrency.setValue(1000)
+    shop.addItem({item: sword, price: goldCurrency})
     let shopItem = shop.getItems()[1]
     return hero.interact(shop).buy(shopItem).then(() => {
+      hero.inventory.increaseCapacity(1)
       expect(hero.inventory.hasItem(shopItem.item)).toEqual(true)
       expect(hero.bank.get({name: 'gold'}).value).toEqual(0)
     })
   })
   test(`should'nt be able to buy item without gold`, () => {
-    shop.addItem(sword)
+    shop.addItem({item: sword, price: goldCurrency})
     let shopItem = shop.getItems()[1]
     return expect(hero.interact(shop).buy(shopItem)).rejects.toEqual(new Error(`Character can't afford`))
   })
@@ -77,10 +79,11 @@ describe('shop', () => {
     hero = factory.hero()
     let armetShopitem = new gg.class.ShopItem({
       item: armet,
+      price: goldCurrency,
       quantity: 1
     })
     return hero.bank.earn(goldCurrency).then(() => {
-      expect(hero.interact(shop).buy(armetShopitem)).rejects.toEqual(new Error('Item not available in shop'))
+      return expect(hero.interact(shop).buy(armetShopitem)).rejects.toEqual(new Error('Item not available in shop'))
     })
   })
 })